Enable feature flag go_user_delete_{branch,tag} by default
These flags have been on at 100% for almost a month now without any
reported issues. See
https://gitlab.com/gitlab-org/gitaly/-/issues/3371 and
https://gitlab.com/gitlab-org/gitaly/-/issues/3370 Let's turn them on
by default.

I was told in a chat with @zj-gitlab that he'd prefer the relevant
operation_service.rb code removed as well, but due to a race condition
on deployment we couldn't do this atomically, it had to be done as a
two-phase rollout.

I.e. as we deploy the Ruby code might be in the middle of
auto-restarting, so we could remove its code before the Go code has a
chance to update with its default, and would still want to call it. So
therefore you need to do any such removal in two gitlab.com release
cycles.