conflicts: Expose helper to merge a conflicting index entry
In order to compute the list of conflicts, we have a function which
takes an `IndexConflict` as input and computes a merged file which
contains the conflicts as part of its merged contents, including
conflict markers. While this function is currently only used in a single
location, we're about to extend the UserMergeToRef RPC to also
optionally commit conflict markers, where this function will be useful.

Let's expose it up front in order to make it available outside of its
current module.