featureflag: Enable Go implementation of FetchSourceBranch by default
While the Go implementation of the FetchSourceBranch RPC was one of the
first ports to happen, it's still not enabled by default. In this case
the reason was not correctness, but performance: as the Go port ommitted
some of the shortcuts which the Ruby port took, it always performed the
fetch and was thus a lot slower than the Ruby code. These performance
shortcomings have been fixed with f382aa958 (Merge branch
'pks-fetch-source-branch-short-circuits' into 'master', 2021-01-08).

While the fixes haven't been deployed that long, the rollout to
production was completely eventless. The port had also been enabled and
tested in the past already without any issues, except for performance.
Considering these two points, it's considered low-risk to default-enable
the feature flag now. So let's do so.