Feature flag rollout doc: remove "update docs?" item
Remove the "Ensure that the documentation has been updated" item. Yes,
docs are good, but this is not actionable. What docs are you going to
update between testing on staging and giving an ETA for production
rollout?

This was just copy/pasted from 2d9d7f9c9 (Add roll out issue template,
2019-06-11), presumably it meant some SRE rollout docs or something
they use. For gitaly the rollout is much simpler, and nobody's
updating any docs at this point in the process.