git: Add new WithEnv() option
There's currently two sets of functions to spawn git commands either
with the default environment or with a supplied environment. In a quest
to reduce the number of variants we have to spawn SafeCmds, this commit
implements a new `WithEnv()` option which can be passed to set up the
environment.