git: Convert needless use of `SafeBareCmd`
The `git.SafeBareCmd()` function serves the purpose to run git commands
which are detached from a repository, e.g. because they don't require
one or because they create one. There's some locations though where
we're running commands which do have an associated repository with this
function.

Let's convert them to use `git.SafeCmd()` instead. This also allows us
to get rid of the manual setup of alternates, which was in fact
forgotten in one of the callsites.