repository: Use `SafeCmdWithEnv()` to check for valid repositories
When calculating the repository checksum, we first check whether the
given repository actually exists on disk and whether it is a valid
repository. We currently do so by using `git rev-parse
--is-inside-git-dir`, which is fine. The downside of it is that we
cannot use `git.SafeBareCmd` with an associated repository as it would
only cause us to pass `--git-dir` to git. Given that it doesn't change
the spawned process's working directory and that `--is-inside-git-dir`
checks whether the current working directory is inside of a git
directory, git would now always pretend to not be inside of a git
directory.

Let's fix the issue by instead switching to `--is-bare-repository`,
which only verifies whether the current repository is bare or not. We're
always going to run this function in bare repositories anyway, and this
switch doesn't require us to change the working directory. So we can now
nicely use `SafeCmdWithEnv()` instead.