git: Rename SafeBareCmdInDir() to NewCommandWithDir()
Back when we introduced the Git DSL, we still had conflicting sets of
safe and unsafe functions. Because of this legacy, our safe set of
functions is still has the "Safe" prefix.

This commit now ends that chapter and renames `SafeBareCmdInDir()`. The
previous name has been quite confusing, given that "bare" in the world
of git typically means a repository without a working directory. That
was not the intended meaning though, but instead that the command is not
run with an associated repository and inside a specific directory. It's
thus renamed to `NewCommandWithDir()`.