FindCommit[s]: add a "Trailers" boolean flag to do %(trailers) work
Amend the code added in 3236114ad (Parse Git commit trailers when
processing commits, 2020-11-26) so we only parse the trailers if it's
requested of us.

This effectively implements a version of what I suggested in the
original MR for it at
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2842#note_461990291

Since then Yorick opened a MR to use this data on the gitaly side:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/49243 as well as
a higher-level UX MR:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/50063

That code clears up the "why do we need to do this work for everything
that uses find_commits" question I had on the original MR, it seems we
really don't, and the combination of this MR and changing the 1st MR
to gitlab.git to have this in app/finders/commits_with_trailer_finder.rb:

    diff --git a/app/finders/commits_with_trailer_finder.rb b/app/finders/commits_with_trailer_finder.rb
    index 5b8c1faec1a..1579b0df35f 100644
    --- a/app/finders/commits_with_trailer_finder.rb
    +++ b/app/finders/commits_with_trailer_finder.rb
    @@ -73,6 +73,6 @@ def each_page(trailer)
       def fetch_commits(offset = 0)
         range = "#{@from}..#{@to}"

    -    @project.repository.commits(range, limit: @per_page, offset: offset)
    +    @project.repository.commits(range, limit: @per_page, offset: offset, trailers: true)
       end
     end

Will save us and /usr/bin/git the work of parsing these out every
time, only for the caller to not care about the data except in this
one case.