git: Rename SafeCmd() to NewCommand()
Back when we introduced the Git DSL, we still had conflicting sets of
safe and unsafe functions. Because of this legacy, our safe set of
functions is still has the "Safe" prefix.

This commit now ends that chapter and renames `SafeCmd()`. In alginment
with the preceding renames, this is being renamed to `NewCommand()`.