Fix ResolveConflicts file limit error
There has been a high occurrence of errors in the Go implementation of
ResolveConflicts in production. After close examination, it was found
that the maximum conflict file size was being inaccurately calculated.
This fix ensures that the file scanner only counts bytes that have
already been advanced, rather than the size of the remaining bytes in
the buffer. This will prevent the same bytes from being counted multuple
times.