git: Rename "unsafe.go" to "command_factory.go"
The "unsafe.go" file has been named that with e5c3c07f5 (git: Rename
files which implement commands, 2021-01-13). This kind of makes sense
given that all functions in there don't do validation of git arguments,
but on another note all functions are receives of the `CommandFactory`
structure.

Let's rename the file again to "command_factory.go". This is in
preparation of upcoming changes where we will drop the unsafe property
altogether.