objectpool: Fix counting of dangling references
When fetching into object pools, we count and log the number of dangling
references we have in an object pool both before and after the fetch.
We do so by using git-for-each-ref(1), enumerating all references in
"refs/dangling". While mostly a theoretical issue, this will miscount
references which start with a different "refs/dangling" prefix like e.g.
"refs/danglinga".

Fix the issue by always appending a trailing slash to the dangling
objects namespace.