git: Merge code paths to generate new commands
The logic to assemble arguments of a new git command is non-trivial and
it's easy to forget some argument. Still we got three different
implementations of this to generate commands with or without a
repository and with or without a directory. This makes the interface
inflexible as it is hard to extend or refactor.

Let's merge the three existing implementations to all call the same
`CommandFactory.newCommand()` function to unify their code paths.