git: Drop unsafe commands
The three wrappers to create unsafe commands are only direct wrappers of
`newCommand()` now. Let's just drop them and call that function
directly. While the current name of `newCommand()` implies that it is
safe, in fact it currently isn't as no argument verification takes
place. This is going to change with the subsequent commit though.