normalize line endings in Go port of UserCommitFiles
WriteBlob currently relies on existing configuration in the repository
to determine whether the line endings should be normalized or not. The
Ruby implementation of UserCommitFiles however always ran with
'core.autocrlf=input'. The Go port has to follow the suit, so this commit
adds and option to normalize the line endings by setting the appropriate
config option on the 'hash-object' invocation. This behavior difference was
caught by a failing test in the QA pipeline of Rails.