git: Convert UpdateRef to accept properly typed arguments
The `UpdateRef()` function acceepts reference as well as its old
respectively new values as plain strings. This makes it easy to miss
what the function is expected to get as input.

Let's convert it to accept a `ReferenceName` as well as two `ObjectID`s
and adjust all callers to pass those instead.