Lint: bring back lint caching
The caching of lints was disabled in b647d26c4 (lint (static code
analysis) enhancements, 2020-04-08). I couldn't see an explanation of
why either in the commit or the MR.

Caching seems to work fine to me, perhaps it was a bug in some older
tool? It's much faster now, running "make lint" in ~1s instead of ~15s
on my box.