Lint: turn "golint" on by default
Since we've got quite a common warning about things needing comments,
add a gigantic list of those exclusions. Generated with:

    make lint |grep golint | sort
    | perl -pe 's/^(.*?):\d+:\d\s*golint\s*(.*)/    - linters: \n        - golint\n      path: "$1"\n      text: "$2"/g'

Except I then had to escape "(" in the text to "\\(" because they're
regexes. I suggested simply turning this specific warning off, but
others in today's team meeting wanted to keep it. So here it is as a
giant list.