operations: Add a testcase for UserCreateTags with transactions
While there's already various tests for UserCreateTags, there's none yet
which verifies that creation of tags works alright in combination with
transactions. Given that we're about to fix two bugs in the code, this
adds a new testcase for this. The testcase demonstrates that
transactions and hooks work alright for the Go code, but fail for the
Ruby code as it would also execute the "update" hook on secondary nodes.