operations: Do not execute update hook on secondaries for Ruby RPCs
When we're about to execute git hooks, then we always need to check
whether the current Gitaly node takes part in a transaction. If so and
if it is not the transaction's primary node, then it mustn't execute
hooks as we'd otherwise execute them multiple times. This always used to
work for the pre-receive and post-receive hooks, but didn't for the
update hook. This was fixed in commit commit 13db4c3c4 (hook: Don't
execute custom update hooks on secondaries, 2020-11-19), but the same
issue exists in our Ruby sidecar.

Fix the issue by correctly passing along transaction information when
executing the update hook.