replace use of server scoped RPC with a repo scoped in TestAuthFailures
With the server scope being removed, this commit removes a use of a server
scoped RPC in tests and replaces it with a repo scoped one. The actual RPC
doesn't matter for the test case as it is testing RPC authentication failures
which are triggered before the handler is reached.