git: Pass Gitaly config when creating new repositories
The `LocalRepository` abstraction indirectly depends on the global
Gitaly configuration as it uses `NewCommand()` to run commands, which in
turn uses the global configuration to resolve repository locations.

As a preparatory step, this commit makes the configuration available to
the `LocalRepository` structure by passing it as an argument of
`NewRepository()`. Note that the configuration is currently not yet
used.