Introduce requestFetchingStartRevision interface
Before UserRevert can revert a commit, it might need to fetch the commit
from a different repository. This is done by the fetchStartRevision
method. This method used the UserRevertRequest to get all info for
fetching that revision.

UserCherryPick needs to do the same. Luckily both request RPCs have the
same fields to do so. This commit changes the input argument to be an
interface that's compatible with both request RPCs. This will allow
UserCherryPick to also use the existing fetchStartRevision method.