Introduction of the CommandFactory interface
The new interface needs to be injected to the places where
the old implementation based on the functions was used.
The interface assembles the high level methods for the commands
creation and hides all required dependencies underneath.

The only implementation of this interface is re-using existing
functions and still rely on the global config.Config var.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699