Substitution of the NewCommandWithDir func with factory method call
The smallest piece we could begin with is to replace usages of the
NewCommandWithDir func with invocation of the NewWithDir method of
the new CommandFactory abstraction. The abstraction implementation
is still relies on the func, but will be refactored to get rid of
config.Config usage.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699