operations: Fix hooks running on secondaries when creating annotated tags
When changing references under a transaction, then we need to make sure
that any hook logic only runs once. This is achieved via the transaction
primary, which gets picked per tranasction and denotes which of the
nodes shall execute hooks. As a result, we always need to pass along
transaction information to where hooks get executed such that we have
info available.

We seemingly already to this when creating annotated tags in our Ruby
codebase, but it turns out that creation of annotated tags would trigger
hooks on each node which is taking part in the transaction. The root
cause is how we call `add_annotated_tag()` at the only callsite. Given
that we pass multiple keyword arguments to it with some of them not
referencing a named argument, they all get accumulated in a single
`options` dict, including our transaction argument. So even though we
thought that we correctly passed along the transaction, it doesn't end
up as the correct argument but instead in the `options` dict.

Fix the issue by simply moving the `transaction` argument before
`options` and not passing it as a keyword. As a result, we now correctly
pass along the transaction and don't invoke hooks on secondaries
anymore.