fix path normalization deviation in UserCommitFiles Go port
The Ruby implementation's path normalization accepts double slashes
in the path as is. This later causes Rugged to error out when attempting
to add a file with such an invalid path to the index.

The Go port is cleaning the filepaths, which replaces double slashes
with single slashes. This makes the filepath valid and thus adding the
file to the index didn't error out.

This commit a special cases double slashes to workaround this quirk.