Failing test for ResolveConflicts with new file
This test modification adds a new commit to each side of a conflict.
These new commits introduce a conflict for a file path that does not
exist in the common ancestor. The Go implementation incorrectly returns
an error indicating it cannot proceed while the Ruby implementation is
able to successfully resolve the conflict. This reproduces an unexpected
error in the GitLab integration tests. See the following issue for
more details:

https://gitlab.com/gitlab-org/gitaly/-/issues/3419