operations: Fix Go port of UserCommitFiles using wrong author/committer
The UserCommitFiles RPC gets as input parameters both the user who is
calling the RPC as well as an optional committer name and email address.
The intention is that the commit's committer is always going to be the
user who is invoking the RPC, while the author can be a different
person.

This works alright for the Ruby code, but the Go code doesn't get this
right. Instead of discerning both author and committer, it will always
use the provided committer name and email in case they are set. If not,
it uses the user's credentials for both.

Fix the issue by correctly discerning both signatures. This commit
doesn't bother to add a testcase as the subsequent commit will do this
already.