conflicts: Allow ResolveConflicts to get an optional timestamp
The ResolveConflicts RPC is currently not deterministic with regards to
the created commits because they contain a timestamp, which is by
default the creation time. This breaks transactions, as when multiple
Gitalies now try to process the same request in parallel, the likelihood
is very high that they'll end up with different commit IDs.

Fix the issue by introducing a new optional "timestamp" field to the
request. If set, it will be used instead of the current system time.
This is only the first half of the fix, the second one will require us
to inject a single timestamp in Rails.