ruby: Add optional parameters to set signature dates
Git signatures for both author and committer include the date on which
the signature has been created. We're currently not really setting this
ourselves in Gitaly, but instead typically use the system's local time.
This is about to change in a quest to make requests more deterministic
and thus workable in transactions, where the same request executed
multiple times must arrive at the same result.

This commit thus includes two ways of adding a specific timestamp to
such signatures to `user_to_committer()` and `User.git_env()`.