Remove Ruby code for old 100% go_user_create_{branch,tag} feature
A follow-up to 0ff3ee285 (Remove on-by-default
go_user_delete_{branch,tag} feature flags, 2021-01-21). Note that the
subject of that commit is a mistake, it's
"go_user_*delete*_{branch,tag}", not "delete".

That commit removed all calls to this code. Since 0ff3ee285 was in
v13.9.0 and we're now starting v13.10.0 we can remove this code. See
the docs I added in 86c8480e8 (Feature flag rollout doc: expand on
post-100% steps, 2021-01-12) for reference.

The s/user_create_branch/user_update_branch/g change in
ruby/spec/lib/gitaly_server/sentry_interceptor_spec.rb is because some
unrelated tests for Sentry depended on calling this method. Let's move
it to another one in the same file to kick that can down the road.

1. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/3033