Makfile: have "lint-strict" also run "lint"
So you don't need to run both, in case the "make lint" target has
issues to report.

Doing it this way means if "lint" has something to report we don't
also run the "lint-strict" target. I experimented with some || -> &&
invocations here to preserve the exit code and always get the full
output, but figured it wasn't worth it.