Makefile: Fix missing dependency for gocover-cobertura tool
The gocover-cobertura tools doesn't currently depend on go.mod, so if
one invokes it in a clean working tree we'd instead use the go.mod of
the parent project, which is Gitaly itself. Let's add the missing
dependency to fix this. This also changes how the test.proto for
Praefect is generated to be more consistent with our other protobuf
files.