Makefile: Move tool binaries into a separate directory
Tool binaries and Gitaly's build artifacts currently live in the same
`_build/bin` directory. This is kind of awkward, as they do have
separate lifetimes and thus shouldn't be treated as the same kind of
artifacts.

This commit makes the separation cleaner by introducing a new
`_build/tools` directory which now hosts all tools required for the
build, only.