Makefile: Create single rule to build external tools
While the instructions to build external build tools is always the same
except for the package, we're using separate rules for all of them. This
makes it hard to refactor the build as one needs to always adjust all
rules at once. Let's fix this by using a pattern rule for our external
tools instead.