repository: Stop pruning if NoPrune was passed to FetchRemote
The FetchRemote RPC accepts a NoPrune parameter, which if set should
cause us to not prune remote references. We fail to pay attention to
that parameter though and always invoke git-fetch(1) with the "--prune"
parameter and inject "remote.$REMOTE.prune=true".

Let's fix this bug by passing the NoPrune parameter to git-fetch(1). We
can also get rid of the injected config entry as it's essentially
duplicating the "--prune" parameter.