repository: Drop remote.$REMOTE.mirror when fetching remotes
When calling FetchRemote with RemoteParams set, we create an ad-hoc
remote and configure it as requested by the RPC's parameters. One of the
config entries we inject at runtime only is "remote.$REMOTE.mirror",
which shouldn't have any effect at all considering it's only used when
pushing to a remote.

Let's drop this useless entry, it's only confusing.