repository: Fix default refspec for FetchRemote
in case a RemoteParam as passed to FetchRemote, we intend do use its
MirrorRefmaps to determine how references should be mirrored, as there
is no configured remote which could have a refspec. In case no map is
provided, we should fall back to just doing a mirror-fetch, if one goes
by what the Ruby RPC did back when it still existed. But in fact we
don't set up a refspec at all in that case, which instead causes us to
simply fetch the remote's HEAD reference.

Fix this issue by using "refs/*:refs/*" in case no refspec is given.