Simplify GoUserUpdateBranch in Go error handling
Instead of checking early if the reference is found we can just
proceed and let update-ref fail. And as noted in [1] and [2] much of
the rest of this can be simplified.

I think I might need to bring some of this back for [3], so I'm doing
this as a commit on top instead of squashing it in.

1. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/3013#note_491544309
2. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/3013#note_491544302
3. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2954