Git command factory injected in to objectpool
In continuation of the removing dependency on the global
config.Config we now replace usages of the git.NewCommand
func with injected factory method designed for the same purpose.
As the calls are deeply nested we are required to make a lot of
changes in other components as well in order to pass the dependency
throw all the layers.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699