testhelper: Use different feature flag to test feature sets
We're using a set of "real" feature flags to test the feature set
implementation. This has the obvious downside that as each feature flag
is eventually being removed, the test cases need to be adjusted to use a
different feature flag instead.

Let's convert this to instead use a set of dummy feature flags.