Makefile: Fix missing plugin path when executing protoc
While three of five of our protoc invocations correctly set the path to
the protoc-gen-go plugin, the remaining ones don't. This works alright
because we have the plugin in our PATH via `_build/bin`, but we're about
to move build tools into a separate "tools" directory.

Let's prepare for this by explicitly specifying the plugin path.