Gitaly transaction service implementation
This commit adds the transaction service implementation for use on
Gitaly servers. Rather than interface with a transaction manager like
the Praefect implementation, the Gitaly transaction service is designed
to only route VoteTransaction and StopTransaction requests via the new
RouteVote RPC.

By keeping the interface for the reference transaction hooks the same,
we minimize the amount of changes in Gitaly. This design shifts the
burden to the integration in the Praefect coordinator.

RouteVote acts as a pub sub message bus, where the Praefect client
subscribes to transactions. Subscribing to a transaction establishes a
route to send messages between Praefect and a Gitaly client.