operations: Wire up AllowConflicts handling for Go
Originally, handling of the "AllowConflicts" parameter for the
UserMergeToRef RPC was only implemented in Ruby and not in Go because
efforts to port to Go and to implement this new flag crossed. This has
changed with the previous commit so that we can now switch over to fully
use the Go port instead of the Ruby code in all cases.

Note that this directly removes the Ruby codepath without any kind of
feature flag. This is done because first, the UserMergeToRef RPC has
been tested already and rolled out by default. And second, all callers
in GitLab Rails which make use of "AllowConflicts" are already hidden
behind the "display_merge_conflicts_in_diff" feature flag which is
currently disabled by default. Introducing another feature flag for this
parameter thus doesn't seem necessary.