housekeeping: Don't treat disappearing files as error
When performing housekeeping tasks, we try to remove various files from
the repository which are stale objects left behind by git. While removal
of these files isn't considered an error, we still log them and
increment the counter of such unremovable files.

Right now, we also treat `ErrNotExist` errors like this. That doesn't
make a whole lot of sense as we know that this file has been removed by
another process now, so this is rather a benign race and not an error.
So let's skip logging of any files which have been removed already.