git: Let LocalRepository implement repository.GitRepo
Eventually, the LocalRepository type shall be prevalent in our codebase
whenever we interact with a git repository. In order to avoid having to
pass around both the protobuf as well as the repository, this commit
converts the LocalRepository to implement the `repository.GitRepo`
interface. This should allow us to pass the LocalRepository instead of
the protobuf in many places.