git: Store LocalRepository in its decendent types
Both LocalRepositoryRemote and LocalRepositoryConfig only store a
`repository.GitRepo` pointer to its LocalRepository parent, which makes
it impossible to use most of the parent's functions. Most importantly,
we want to be able to run commands on the parent without going through
the `git.NewCommand()` interface.

Fix this by storing a pointer to `LocalRepository` instead.