CI: retry to work around flaky tests
We have flaky tests[1] that keep causing false alarms in the CI, so
people have to waste time looking at the same issue and click "retry"
manually.

Let's do that retry automatically. We'll waste less time just having
it retry so you'll potentially wait a bit more for a real failure,
v.s. the false alarms.

This is an interim solution for[1]. We should also fix those tests, a
smarter way to do this would be to quickly annotate known failed
tests, or to have "make test" et al smartly retry just the failing
tests, but just doing this is simpler and better than the status quo.

1. https://gitlab.com/gitlab-org/gitaly/-/issues?label_name[]=failure%3A%3Aflaky-test