Injection of the git.CommandFactory into catfile package
As we are abstracting dependencies required to create a git
command we introduced git.CommandFactory interface that is
in duty of creation of the git commands. The catfile package
used across different services and as a first step we provide
factory as an additional constructor parameter.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699