hook: Extract transaction handling into standalone manager
The hook manager currently hosts all logic to handle transactional
voting. This made sense until now as the reference-transaction hook was
the only place where we would do voting on transactions in the first
place. But we're about to extend transactional support to more places
which are not related to references at all.

As a preparatory step, this commit thus extracts the transaction
handling code into its own transaction manager. To keep the diff small
and focussed on the functional change, the manager is still created
ad-hoc in the hook manager. This is about to change with a subsequent
commit.