repository: Use transactions when writing gitattributes
The `ApplyGitattributes()` RPC reads an optional ".gitattributes" file
from a given revision and writes it into "info/attributes". Like this,
it's made sure that the the gitattributes file from the repo's main
branch correctly influences operations on the git repository, e.g. when
using the Web IDE.

As the current scope of transactions was only about updating git
references in a consistent way, the gitattributes file won't use them at
all as no reference is being updated. This does have its downsides
though: the file is written as it is triggered whenever the main branch
gets updated, and not using transactions now forces us to create
replication jobs as repository state has been changed. This is wasteful
and also leads to repositories being out-of-date quite often.

This commit thus introduces support for transactions when updating the
gitattributes file. While transactions were originally catered to
reference updates only, the way they're implemented is completely
agnostic of what's being voted on: all that is required is a unique hash
to vote on. And we've already got that, as we can simply take the blob
ID of the gitattributes file which we're about to write, whereas we take
the zero OID when it's about to be removed.