transactions: Log transaction ID if transaction was not found
When a node casts a vote on a transaction, then it needs to pass the
identifier of the transaction in order to identify it. If it is not
possible to find a transaction for this ID, then we have to return an
error.

This error doesn't currently include the transaction ID, which makes it
quite hard to correlate the error with other ongoing requests. Fix this
by including the ID in the error message.