transactions: Discern cancellation and failure
When a transaction gets cancelled or when the transaction failed to
reach quorum, then we both account this under the `VoteAborted` result.
It is thus impossible to tell apart both cases, which is especially bad
when one tries to comprehend logs.

This commit thus splits the state into two, `VoteFailed` which
represents that the vote did not reach majority. And `VoteCanceled`,
which represents that the vote was cancelled before majority was
reached.