Replacement of the git.NewCommandWithoutRepo in repository package
Replacement of the git.NewCommandWithoutRepo with the
git.CommandFactory.NewWithoutRepo call and git.NewCommand with
git.CommandFactory.New. The functions were migrated to methods
in order to get access to the git.CommandFactory without need to
add it as an additional parameter.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699